Add files via upload
The event map fills the 10 by 10 grid with values. Each turn, the event check function is used to take the current position of the player,
and check if that tile has an associated event that needs to occur. It then calls the apropriate function. Needs some more events added to the grid, is just an early version. Needs to be incorporated into the main code.